corpus-analyzer/pyproject.toml
[project]
name = "corpus-analyzer"
version = "0.1.0"
description = "Extract, categorize, analyze, and template documentation from folders of docs and scripts"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer[all]>=0.12.0",
    "rich>=13.7.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "sqlite-utils>=3.36",
    "python-frontmatter>=1.1.0",
    "ollama>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.4.0",
    "mypy>=1.9.0",
]

[project.scripts]
corpus-analyzer = "corpus_analyzer.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/corpus_analyzer"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
